backend/pyproject.toml
[tool.poetry]
name = "dns-max-backend"
version = "0.1.0"
description = "Domain reseller system backend API"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.29.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "^42.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.25.2"
tencentcloud-sdk-python-dnspod = "^3.0.1043"
alibabacloud-alidns20150109 = "^3.0.1"
alibabacloud-tea-openapi = "^0.3.9"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"